refactor: add `NotificationsWithLinks` react context
